refactor: using userId from token for update and delete User, Prodcut and offer

+++ src/main/java/com/martelando/martelandoapp/sevice/impl/OfferServiceImpl.java
@@ -4,10 +4,13 @@ import com.martelando.martelandoapp.controllers.request.SaveOfferRequest;
 import com.martelando.martelandoapp.controllers.request.UpdateOfferRequest;
 import com.martelando.martelandoapp.controllers.responses.OfferDetailResponse;
 import com.martelando.martelandoapp.exception.NotFoundException;
+import com.martelando.martelandoapp.exception.UnauthorizedOfferUpdate;
 import com.martelando.martelandoapp.exception.UserAlreadyMakeOfferException;
 import com.martelando.martelandoapp.mapper.IOfferMapper;
 import com.martelando.martelandoapp.repository.IOfferRepository;
+import com.martelando.martelandoapp.repository.IUserRepository;
 import com.martelando.martelandoapp.sevice.IOfferService;
+import com.martelando.martelandoapp.sevice.IUserService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,14 +22,21 @@ public class OfferServiceImpl implements IOfferService {
 
     private final IOfferRepository offerRepository;
     private final IOfferMapper offerMapper;
+    private final IUserRepository userRepository;
+
 
     @Override
-    public OfferDetailResponse create(SaveOfferRequest saveOfferRequest) {
+    public OfferDetailResponse create(Long userId, SaveOfferRequest saveOfferRequest) {
         var offer = this.offerRepository.findByBidderIdAndProductId(saveOfferRequest.bidderId(), saveOfferRequest.productId());
+
+        if(!userRepository.existsById(userId)) {
+            throw new NotFoundException("Usuario não existe");
+        }
 
         if(offer.isPresent()) {
             throw new UserAlreadyMakeOfferException("Usuario já fez uma oferta para esse produto");
         }
+
 
         var offerSaved = this.offerRepository.save(
                 this.offerMapper.toEntity(saveOfferRequest)
@@ -36,17 +46,25 @@ public class OfferServiceImpl implements IOfferService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long userId, Long id) {
         var offer = this.offerRepository.findById(id)
                 .orElseThrow(() ->  new NotFoundException("Oferta não existe"));
+
+        if(!offer.getBidder().getId().equals(userId)) {
+            throw new UnauthorizedOfferUpdate("Usuario não autorizado a deletar oferta");
+        }
 
         this.offerRepository.delete(offer);
     }
 
     @Override
-    public OfferDetailResponse update(UpdateOfferRequest updateOfferRequest) {
+    public OfferDetailResponse update(Long userId,UpdateOfferRequest updateOfferRequest) {
         var offer = this.offerRepository.findById(updateOfferRequest.id())
                 .orElseThrow(() -> new NotFoundException("Oferta não existe"));
+
+        if(!offer.getBidder().getId().equals(userId)) {
+            throw new UnauthorizedOfferUpdate("Usuario não autorizado a deletar oferta");
+        }
 
         offer.setStatus(updateOfferRequest.status());
         offer.setAmount(updateOfferRequest.amount());
